Scala update to 2.12.2. Akka updates. meta-core update.
Akka HTTP updated to 10.0.9.
Akka updated to 2.4.19.
DataObject from meta-core now has fileName as String, not
Option[String].

+++ src/main/scala/se/lu/nateko/cp/data/formats/DateColTimeColTimeSeriesToBinTableConverter.scala
@@ -30,17 +30,17 @@ abstract class DateColTimeColTimeSeriesToBinTableConverter(colFormats: ColumnFor
 
 		if(date == dateNull || time == timeNull) stampNull
 		else{
-			val locDate = LocalDate.ofEpochDay(date)
+			val locDate = LocalDate.ofEpochDay(date.toLong)
 
 			val dt =
 				if(time >= 86400){
-					val locTime = LocalTime.ofSecondOfDay(time - 86400)
-					LocalDateTime.of(locDate, locTime).plusHours(24 - offsetFromUtc)
+					val locTime = LocalTime.ofSecondOfDay((time - 86400).toLong)
+					LocalDateTime.of(locDate, locTime).plusHours((24 - offsetFromUtc).toLong)
 				} else {
-					val locTime = LocalTime.ofSecondOfDay(time)
-					LocalDateTime.of(locDate, locTime).minusHours(offsetFromUtc)
+					val locTime = LocalTime.ofSecondOfDay(time.toLong)
+					LocalDateTime.of(locDate, locTime).minusHours(offsetFromUtc.toLong)
 				}
-			Double.box(dt.toInstant(ZoneOffset.UTC).toEpochMilli)
+			Double.box(dt.toInstant(ZoneOffset.UTC).toEpochMilli.toDouble)
 		}
 	}
